+++ packages/effect-http-node/examples/multiple-responses.ts
@@ -1,10 +1,8 @@
 import { NodeRuntime } from "@effect/platform-node"
 import { Schema } from "@effect/schema"
-import { Effect, pipe } from "effect"
+import { Effect, Logger, LogLevel, pipe, Random } from "effect"
 import { Api, ApiResponse, RouterBuilder } from "effect-http"
-
 import { NodeServer } from "effect-http-node"
-import { debugLogger } from "./_utils.js"
 
 const helloEndpoint = Api.post("hello", "/hello").pipe(
   Api.setResponseBody(Schema.Number),
@@ -23,15 +21,24 @@ const api = pipe(
   Api.addEndpoint(helloEndpoint)
 )
 
+const randomChoice = <A extends ReadonlyArray<any>>(...values: A) =>
+  Random.nextIntBetween(0, values.length).pipe(Effect.map((i) => values[i]))
+
 const app = pipe(
   RouterBuilder.make(api),
-  RouterBuilder.handle("hello", () => Effect.succeed({ body: 12, headers: { "my-header": 69 }, status: 201 as const })),
+  RouterBuilder.handle("hello", () =>
+    randomChoice(
+      { status: 200, body: 12, headers: { "my-header": 69 } },
+      { status: 201, body: 12 },
+      { status: 204, headers: { "x-another": 12 } }
+    )),
   RouterBuilder.build
 )
 
 pipe(
   app,
   NodeServer.listen({ port: 3000 }),
-  Effect.provide(debugLogger),
+  Effect.provide(Logger.pretty),
+  Logger.withMinimumLogLevel(LogLevel.All),
   NodeRuntime.runMain
 )