+++ packages/effect-http/src/Handler.ts
@@ -6,15 +6,16 @@
 import type * as HttpRouter from "@effect/platform/HttpRouter"
 import type * as AST from "@effect/schema/AST"
 import type * as Effect from "effect/Effect"
+import type * as Pipeable from "effect/Pipeable"
 import type * as Types from "effect/Types"
 
-import type * as Security from "effect-http-security/Security"
-import type * as Pipeable from "effect/Pipeable"
 import type * as ApiEndpoint from "./ApiEndpoint.js"
 import type * as ApiRequest from "./ApiRequest.js"
 import type * as ApiResponse from "./ApiResponse.js"
+import type * as HttpError from "./HttpError.js"
 import * as internal from "./internal/handler.js"
 import type * as utils from "./internal/utils.js"
+import type * as Security from "./Security.js"
 
 /**
  * @category type id
@@ -123,7 +124,31 @@ export declare namespace Handler {
    * @since 1.0.0
    */
   export type ToSecurity<Security extends Security.Security.Any> = Security.Security.Success<Security>
+
+  /**
+   * @category models
+   * @since 1.0.0
+   */
+  export type Endpoint<H> = H extends Handler<infer A, any, any> ? A : never
+
+  /**
+   * @category models
+   * @since 1.0.0
+   */
+  export type Error<H> = H extends Handler<any, infer E, any> ? E : never
+
+  /**
+   * @category models
+   * @since 1.0.0
+   */
+  export type Context<H> = H extends Handler<any, any, infer R> ? R : never
 }
+
+/**
+ * @category constructors
+ * @since 1.0.0
+ */
+export const empty: Handler<never, never, never> = internal.empty
 
 /**
  * @category constructors
@@ -133,12 +158,12 @@ export const make: {
   <A extends ApiEndpoint.ApiEndpoint.Any, E, R>(
     fn: Handler.Function<A, E, R>,
     options?: Partial<Options>
-  ): (endpoint: A) => Handler<A, E, R>
+  ): (endpoint: A) => Handler<A, Exclude<E, HttpError.HttpError>, R>
   <A extends ApiEndpoint.ApiEndpoint.Any, E, R>(
     endpoint: A,
     fn: Handler.Function<A, E, R>,
     options?: Partial<Options>
-  ): Handler<A, E, R>
+  ): Handler<A, Exclude<E, HttpError.HttpError>, R>
 } = internal.make
 
 /**
@@ -156,17 +181,40 @@ export const makeRaw: {
 } = internal.makeRaw
 
 /**
- * @category getters
+ * @category combinators
  * @since 1.0.0
  */
-export const getRoute: <A extends ApiEndpoint.ApiEndpoint.Any, E, R>(
-  handler: Handler<A, E, R>
-) => HttpRouter.Route<E, R> = internal.getRoute
+export const concat: {
+  <A extends ApiEndpoint.ApiEndpoint.Any, B extends ApiEndpoint.ApiEndpoint.Any, E1, E2, R1, R2>(
+    self: Handler<A, E1, R1>,
+    handler: Handler<B, E2, R2>
+  ): Handler<A | B, E1 | E2, R1 | R2>
+
+  <B extends ApiEndpoint.ApiEndpoint.Any, E2, R2>(
+    handler: Handler<B, E2, R2>
+  ): <A extends ApiEndpoint.ApiEndpoint.Any, E1, R1>(self: Handler<A, E1, R1>) => Handler<A | B, E1 | E2, R1 | R2>
+} = internal.concat
 
 /**
- * @category getters
+ * @category combinators
  * @since 1.0.0
  */
-export const getEndpoint: <A extends ApiEndpoint.ApiEndpoint.Any, E, R>(
+export const concatAll: <Handlers extends ReadonlyArray<Handler.Any>>(
+  ...handlers: Handlers
+) => Handler<Handler.Endpoint<Handlers[number]>, Handler.Error<Handlers[number]>, Handler.Context<Handlers[number]>> =
+  internal.concatAll
+/**
+ * @category destructors
+ * @since 1.0.0
+ */
+export const getRouter: <A extends ApiEndpoint.ApiEndpoint.Any, E, R>(
   handler: Handler<A, E, R>
-) => A = internal.getEndpoint
+) => HttpRouter.HttpRouter<E, R> = internal.getRouter
+
+/**
+ * @category destructors
+ * @since 1.0.0
+ */
+export const getEndpoints: <A extends ApiEndpoint.ApiEndpoint.Any, E, R>(
+  handler: Handler<A, E, R>
+) => ReadonlyArray<A> = internal.getEndpoints